task 6: add storing in db of passed product from csv
