+++ product-service/src/functions/catalogBatchProcess/handler.ts
@@ -1,23 +1,37 @@
 import 'source-map-support/register';
 
 import { middyfy } from '@libs/lambda';
+import { formatJSONError, formatJSONResponse } from '@libs/api-gateway';
+
 import { DbConnectService } from '@services/db-connect.service';
 import { SnsManagementService } from '@services/sns-management.service';
-// import {ProductService} from "@services/product.service";
-// import { ProductService } from '@services/product.service';
+import { ProductService } from '@services/product.service';
 
-const catalogBatchProcess = (snsManagementService: SnsManagementService) => async (event) => {
-  for (const { body } of event.Records) {
-    const { price } = JSON.parse(body);
+const catalogBatchProcess = (productService: ProductService, snsManagementService: SnsManagementService) => async (event) => {
+  try {
+    for (const { body } of event.Records) {
+      const product = JSON.parse(body);
+      console.log(`Message body: ${body}`);
 
-    console.log(`Message body: ${body}`);
-    await snsManagementService.publish(body, price);
-    console.log('Publishing to SNS is finished');
+      if (!productService.isProductValid(product)) {
+        return formatJSONError({ message: 'Please pass correct product structure' }, 400);
+      }
+
+      const createdProduct = await productService.createProduct(product);
+      console.log(`Product is stored in db: ${JSON.stringify(createdProduct)}`);
+
+      await snsManagementService.publish(body, product.price);
+      console.log('Publishing to SNS is finished');
+    }
+
+    return formatJSONResponse();
+  } catch (err) {
+    return formatJSONError({ message: err });
   }
 }
 
 const dbConnectService = new DbConnectService();
 const snsManagementService = new SnsManagementService();
-// const productService = new ProductService(dbConnectService);
+const productService = new ProductService(dbConnectService);
 
-export const main = middyfy({ handler: catalogBatchProcess(snsManagementService), dbConnectService });
+export const main = middyfy({ handler: catalogBatchProcess(productService, snsManagementService), dbConnectService });
